backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
supabase==2.3.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
joblib==1.3.2
paho-mqtt==1.6.1
reportlab==4.0.7
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.24.1
kagglehub==0.2.1
requests==2.31.0
matplotlib==3.8.2
plotly==5.18.0
pillow==10.1.0
python-pptx==0.6.23

